Refactor xrefs to general references processor
Ahead of work to include things like glossaries and bibliographies (in a
"jupyter book" way), this is a minor refactor/renaming to better
facilitate that work in the future, allowing us to have a
multi-reference processing file to hold more kinds of functions.
