+++ jupyter_book_to_htmlbook/reference_processing.py
